Support KGMS 1.5.2 (#42)
## What is the goal of this PR?

**NOTE**: This change is an internal change, the public API stays the same.

Currently, `client-python` only supports KGMS 1.4.3 and won't work with 1.5. This is because in 1.5 the Session / Transaction architecture were simplified and `username` and `password` needs to be correctly supplied when opening a session rather than transaction. This PR fixes that.

## What are the changes implemented in this PR?
When opening a session, `username` and `password` are supplied in the `Session.Open.Req` gRPC message.