Delete collect_concepts (#77)
## What is the goal of this PR?
Shortcut method `answer_iterator.collect_concepts()` was leading to user confusion as it can lead to duplicate concepts being returned in the same list. As it doesn't really fit with the general paradigm of simplifying and removing syntactic sugar, we delete this method and ask users to replace calls with a piece of custom code for now. 
Replacement is very easy when `collect_concepts` was associated with a query with 1 variable:
`[ans.get("x") for ans in answer_iterator]`


## What are the changes implemented in this PR?
* Delete `collect_concepts` and associated usage in tests