Fix leaking GRPC threads on transaction error (#114)
## What is the goal of this PR?

We block the GRPC request observer in order to wait for new client requests. If the transaction errors and we do not unblock this observer, the GRPC thread will be left waiting forever. Previously, we had an error case that would result in this, so this PR patches that case by allowing `close()` to work correctly even on an error.

## What are the changes implemented in this PR?

- Stopped the communicator becoming closed on error.
- Ensure the communicator can close even if the transaction errored.