Refactor 2.0 build (#124)
## What is the goal of this PR?
Update to use cleaned up deployment targets, new format of deployment properties in `deployment.bzl`. We also implement the new style of running release-dependency validation using a bazel test rule instead of a script called in CI. This required adding `workspace_refs` to WORKSPACE.

NOTE: the current state is that client-python is not expected to run or pass tests, as Grakn 2.0 is not finished, and client-python is not updated at all yet with 2.0 changes in protocol etc.