Add BDD to CI and create pip deployment test (#157)
## What is the goal of this PR?

Our test suite was not being run in CI and could only be triggered manually by ourselves. In order to get them to work in CI, we needed them to fetch, unzip and run Grakn in the background. So we implemented that and added the tests to CI.

We also didn't have an automated mechanism for verifying that our `pip` deployment produced a working Client Python artifact that could be used to make Graql queries, so we created `test-deployment-pip`.

## What are the changes implemented in this PR?

- Unzip and run Grakn when running BDD tests
- Add test-behaviour to automation.yml
- Create test-deployment-pip