Grakn Cluster integration (#171)
## What is the goal of this PR?

We now support connecting to Grakn Cluster.

This is a **breaking change.** Creating a `GraknClient` now requires a new syntax:
```
client = GraknClient.core(address)
```
To connect to Grakn Cluster, use
```
client = GraknClient.cluster(address)
```
The API usage of a Cluster client is the same as for a Core client. It also supports one additional field in `GraknOptions` named `allow_secondary_replica`, which when set to `True` allows **read** transactions to read from secondary database replicas.

## What are the changes implemented in this PR?

Create Cluster-specific versions of GraknClient, Session and DatabaseManager for use when connecting to Cluster.