When testing Grakn, randomize its port and poll to detect its startup (#180)
## What is the goal of this PR?

Previously when running tests, Grakn would always start on port 1729. This rendered it impossible for us to parallelise our tests. Now, we pick a random port number between 40000 and 60000.

Also, previously we waited N milliseconds and then ran the tests. This was wasteful, because often Grakn would start way in advance; and unreliable, because sometimes Grakn would not start in time and CI would fail. Now, we add a polling loop that checks if the server port is in use, and signals that the tests may begin once the port is in use.

## What are the changes implemented in this PR?

- Randomize port when running Grakn during BDD tests
- During BDD tests, poll the ports to detect when Grakn has actually started
- Enable test parallelisation in CI