Depend on grpcio<2, protobuf<4 (#275)
## What is the goal of this PR?

We increased the flexibility of our dependency requirements. Any version
of `grpcio < 2` and `protobuf < 4` are now acceptable.

## What are the changes implemented in this PR?

We have previously had to do a new release of this client whenever
someone depends on a different version of `grpcio` or `protobuf` in
their project - in particular, even if they are depending on a _newer_
version. This is not a good experience for developers.

Now, we accept any version that is _expected_ to be non-breaking.